requirements.txt
playwright>=1.45.0
PyYAML>=6.0.1
packaging>=23.0
msal>=1.28.0
requests>=2.31.0